Abstract changelog logic into hook
